adds support for overwriting file storage backends
This change modifies the admin edit view for both documents and images
to prevent deletion of uploaded files when the storage backend
implements file overwriting.

See #4495 and #4502 as examples of previously reported issues where this
happened with the django-storages S3 backends, which use overwriting by
default.

The current behavior (implemented in 1.13 in #3821 and #3822) is to
delete an old file when a new file is uploaded for a document or image.
Normally this works fine, as Django file storage backends are supposed
to give each upload a unique name (typically by appending random
characters to filenames).

However, there are some backends that support overwriting of files
instead, meaning that a file that gets uploaded with the same name as an
existing file will overwrite that file, instead of getting stored
alongside it. In these cases, the current Wagtail logic first overwrites
the file and then deletes it, leaving a broken document/image with no
stored file.

This logic change simply checks to make sure that the old name and the
new name are in fact different; if the backend assigned them the same
name then the deletion is skipped. This shouldn't have any downside but
does allow for continued use of overwriting backends.

This change includes a new unit test that checks this behavior.

This change also reverts part of the documentation added in #4528, as
using  django-storages S3 backends in their default configurations (with
AWS_S3_FILE_OVERWRITE set to True) is now possible.

+++ wagtail/documents/views/documents.py
@@ -152,7 +152,15 @@ def edit(request, document_id):
                 # if providing a new document file, delete the old one.
                 # NB Doing this via original_file.delete() clears the file field,
                 # which definitely isn't what we want...
-                original_file.storage.delete(original_file.name)
+                #
+                # Only do this if the new stored file was given a different
+                # filename than the old one. By convention Django file storage
+                # does not overwrite uploaded files that have the same name,
+                # but some storage backends do support this. This extra check
+                # ensures that if a storage backend overwrites an existing
+                # file, we don't delete what we just uploaded.
+                if doc.file.name != original_file.name:
+                    original_file.storage.delete(original_file.name)
 
             # Reindex the document to make sure all tags are indexed
             search_index.insert_or_update_object(doc)
